+++ classes/local/grading/kany.php
@@ -13,32 +13,33 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+namespace qtype_matrix\local\grading;
+
+use qtype_matrix\local\qtype_matrix_grading;
 
 /**
+ * Per row grading. The total grade is the average of grading received
+ * for reach one of the rows.
  *
- * Total grading - i.e. including rows.
- *
- * The student must choose all correct answers, and none of the wrong ones
- * to get 100% otherwise he gets 0%. Including rows.
- * If one row is wrong then the mark for the question is 0.
+ * Any correct and no wrong answer to get 100% otherwise 0
  */
-class qtype_matrix_grading_kprime extends qtype_matrix_grading {
+class kany extends qtype_matrix_grading {
 
-    const TYPE = 'kprime';
+    const TYPE = 'kany';
 
     public static function get_name() {
         return self::TYPE;
     }
 
     public static function get_title() {
-        return qtype_matrix::get_string(self::TYPE);
+        return \qtype_matrix::get_string(self::TYPE);
     }
 
     /**
      * Factory
      *
      * @param string $type
-     * @return qtype_matrix_grading_kprime
+     * @return kany
      */
     public static function create($type) {
         static $result = false;
@@ -48,41 +49,47 @@ class qtype_matrix_grading_kprime extends qtype_matrix_grading {
         return $result = new self();
     }
 
-    /**
-     * Returns the question's grade. By default it is the average of correct questions.
-     *
-     * @param qtype_matrix_question $question
-     * @param array                 $answers
-     * @return float
-     */
     public function grade_question($question, $answers) {
+        $numberofcorrectrows = 0;
         foreach ($question->rows as $row) {
             $grade = $this->grade_row($question, $row, $answers);
-            if ($grade < 1) {
-                return 0;
+            if ($grade >= 1) {
+                $numberofcorrectrows++;
             }
         }
-        return 1;
+        if ($numberofcorrectrows == count($question->rows)) {
+            return 1;
+        } else if ((count($question->rows) - $numberofcorrectrows) == 1) {
+            return 0.5;
+        }
+        return 0;
     }
-
 
     /**
      * Grade a row
      *
-     * @param qtype_matrix_question $question  The question to grade
+     * @param \qtype_matrix_question $question  The question to grade
      * @param integer|object        $row       Row to grade
      * @param array                 $responses User's responses
      * @return float                            The row grade, either 0 or 1
      */
     public function grade_row($question, $row, $responses) {
+        $onecorrectanswer = false;
         foreach ($question->cols as $col) {
             $answer = $question->answer($row, $col);
             $response = $question->response($responses, $row, $col);
-            if ($answer != $response) {
+            if (!$answer && $response) {
                 return 0;
             }
+            if ($answer && $response) {
+                $onecorrectanswer = true;
+            }
         }
-        return 1;
+        return ($onecorrectanswer) ? 1 : 0;
+    }
+
+    public function validation($data) {
+        return [];
     }
 
 }
